Fix loading/saving options
