+++ CurePlease.Model/JobUtils.cs
@@ -1,0 +1,141 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CurePlease.Model
+{
+    public static class JobUtils
+    {
+        public class JobTitles : List<JobTitles>
+        {
+            public int job_number
+            {
+                get; set;
+            }
+
+            public string job_name
+            {
+                get; set;
+            }
+        }
+
+        public static List<JobTitles> JobNames = new List<JobTitles>();
+
+        static JobUtils()
+        {
+            JobNames.Add(new JobTitles
+            {
+                job_number = 1,
+                job_name = "WAR",
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 2,
+                job_name = "MNK"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 3,
+                job_name = "WHM"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 4,
+                job_name = "BLM"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 5,
+                job_name = "RDM"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 6,
+                job_name = "THF"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 7,
+                job_name = "PLD"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 8,
+                job_name = "DRK"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 9,
+                job_name = "BST"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 10,
+                job_name = "BRD"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 11,
+                job_name = "RNG"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 12,
+                job_name = "SAM"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 13,
+                job_name = "NIN"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 14,
+                job_name = "DRG"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 15,
+                job_name = "SMN"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 16,
+                job_name = "BLU"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 17,
+                job_name = "COR"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 18,
+                job_name = "PUP"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 19,
+                job_name = "DNC"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 20,
+                job_name = "SCH"
+            });
+
+            JobNames.Add(new JobTitles
+            {
+                job_number = 21,
+                job_name = "GEO"
+            });
+            JobNames.Add(new JobTitles
+            {
+                job_number = 22,
+                job_name = "RUN"
+            });
+        }
+    }
+}
